chore: move some helpers to own util file
