+++ scripts/utils/mdast.js
@@ -1,0 +1,8 @@
+const yaml = require('js-yaml');
+
+const frontmatter = (attributes) => ({
+  type: 'yaml',
+  value: yaml.safeDump(attributes),
+});
+
+module.exports = { frontmatter };
